[WIP] Update core modules.

+++ build/Commands/Setup.js
@@ -25,16 +25,38 @@ class Setup extends Cli_1.CommandDefines {
     async initialize() {
         this.cli
             .command('app setup', 'Install and configure components.')
+            .option('-f, --force', 'Forced reconfiguration of components.')
+            .option('-e, --env', 'Save as current environment settings.')
+            .option('-y, --yes', 'Assume yes if prompted.')
             .action((args) => {
             return new Promise(async (resolve, reject) => {
+                /*  */
+                if (args.options.force && !args.options.yes) {
+                    let forcePrompt = await this.cli.activeCommand.prompt({
+                        type: 'confirm',
+                        name: 'continue',
+                        default: false,
+                        message: 'This action can break the application. Continue?',
+                    });
+                    if (!forcePrompt.continue) {
+                        this.logger.info('Canceled by user.');
+                        return resolve();
+                    }
+                }
+                /*  */
+                const profiler = this.logger.startTimer();
+                /*  */
                 let list = [];
+                /*  */
                 list.push(async (command, args) => {
                     /* Check and create boot config file. */
                     if (!fs_1.default.existsSync(Const_1.BOOT_FILE)) {
                         fs_1.default.writeFileSync(Const_1.BOOT_FILE, JSON.stringify({}));
                     }
                 });
+                /*  */
                 this.events.emit('app:getSetupSubscriptions', list);
+                /*  */
                 for (const task of list) {
                     if (typeof task === 'function') {
                         try {
@@ -45,6 +67,10 @@ class Setup extends Cli_1.CommandDefines {
                         }
                     }
                 }
+                /*  */
+                profiler.done({
+                    message: 'Install and configure components completed.'
+                });
                 resolve();
             });
         });
